test: moved both test suites into one file
